+++ Assets/Scenes/SampleScene 2_1.unity
@@ -960,6 +960,14 @@ PrefabInstance:
       propertyPath: m_LinearDrag
       value: 5
       objectReference: {fileID: 0}
+    - target: {fileID: 7991418579374025228, guid: 92d9412970dea7a4a8315d9912d22d3b, type: 3}
+      propertyPath: rb
+      value: 
+      objectReference: {fileID: 1309868729}
+    - target: {fileID: 7991418579374025228, guid: 92d9412970dea7a4a8315d9912d22d3b, type: 3}
+      propertyPath: cam
+      value: 
+      objectReference: {fileID: 519420031}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 92d9412970dea7a4a8315d9912d22d3b, type: 3}
 --- !u!1 &1309868717 stripped
@@ -970,6 +978,11 @@ GameObject:
 --- !u!4 &1309868726 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 217788455370571802, guid: 92d9412970dea7a4a8315d9912d22d3b, type: 3}
+  m_PrefabInstance: {fileID: 1309868716}
+  m_PrefabAsset: {fileID: 0}
+--- !u!50 &1309868729 stripped
+Rigidbody2D:
+  m_CorrespondingSourceObject: {fileID: 2103830374191505694, guid: 92d9412970dea7a4a8315d9912d22d3b, type: 3}
   m_PrefabInstance: {fileID: 1309868716}
   m_PrefabAsset: {fileID: 0}
 --- !u!1 &1509620277
