+++ Assets/Scripts/DeleteMe.cs
@@ -5,7 +5,7 @@ using static Vector;
 using static Func;
 using static State;
 using static Event;
-public class PlayerController : MonoBehaviour, IDamageSource
+public class DeleteMe : MonoBehaviour, IDamageSource
 {
     float _munch = 1;
 
@@ -13,7 +13,7 @@ public class PlayerController : MonoBehaviour, IDamageSource
     {
         get
         {
-            var vec5 = new Vector3(5, 5, 5);
+            var vec5 = new Vector3(5,5,5);
             _munch += 0.01f;
             return Vector3.Min(vec5, transform.localScale * _munch);
         }
@@ -50,7 +50,6 @@ public class PlayerController : MonoBehaviour, IDamageSource
 
     public float DashMod = 125f;
 
-
     private void Awake()
     {
         rb = rb is null ? GetComponent<Rigidbody2D>() : rb;
@@ -65,12 +64,12 @@ public class PlayerController : MonoBehaviour, IDamageSource
 
         static float dst_origin(Vector2 vec2) => Compose(Mathf.Abs, Vector.Distance(vec2))(Vector2.zero);
 
-        var move_state = (Move: dst_origin(movement),
+        var move_state = (Move:     dst_origin(movement),
                           Velocity: dst_origin(rb.velocity));
 
         @event = move_state switch
         {
-            { Move: 0, Velocity: <= 0.1f } => @event.ExceptFor(DoWalk | DoMadDash).Union(EventsFromKeys()),
+            { Move: 0, Velocity: <= 0.1f } => @event.ExceptFor(DoWalk|DoMadDash).Union(EventsFromKeys()),
 
             { Move: not 0 } => @event.Union(DoWalk).ExceptFor(DoMadDash),
             { Move: 0 } => @event.ExceptFor(DoWalk).Union(EventsFromKeys()),
@@ -108,14 +107,10 @@ public class PlayerController : MonoBehaviour, IDamageSource
     void OnTriggerEnter2D(Collider2D col)
     {
         Debug.Log(col.name);
-        if (!state.HasFlag(MadDashing))
+        if (col.CompareTag("Enemy") || col.CompareTag("ABullet") && (timer - lastHitSnap) > 2.0)
         {
-
-            if (col.CompareTag("Enemy") || col.CompareTag("ABullet") && (timer - lastHitSnap) > 2.0)
-        {
-                lastHitSnap = timer;
-                life--;
-            }
+            lastHitSnap = timer;
+            life--;
         }
         else if (@state.HasFlag(MadDashing))
         {
@@ -151,19 +146,6 @@ public class PlayerController : MonoBehaviour, IDamageSource
         {
             BaseEnemyController enemyCtr = col.GetComponent<BaseEnemyController>();
             enemyCtr.TakeDamage(damage);
-        }
-    }
-
-    public void TakeDamage(int damage,string tag)
-    {
-        if (!state.HasFlag(MadDashing))
-        {
-
-            if (tag == "Enemy" || tag == "ABullet"  && (timer - lastHitSnap) > 2.0)
-            {
-                lastHitSnap = timer;
-                life--;
-            }
         }
     }
 }
