+++ Assets/Scripts/DeleteMe.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b4e870b2f20ae9746961ea0cab999d9b
+guid: 0d9137f7f6d7c96429ea2d8a21edd7fa
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
